requirements.txt
asgiref==3.9.2
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.3.0
constantly==23.10.4
cryptography==46.0.1
daphne==4.2.1
defusedxml==0.7.1
Django==5.2.6
django-allauth==65.11.2
django-widget-tweaks==1.5.0
fonttools==4.60.0
fpdf2==2.8.4
geographiclib==2.1
geopy==2.4.1
gunicorn==23.0.0
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
redis==6.4.0
requests==2.32.5
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
stripe==12.5.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
zope.interface==8.0
whitenoise
